pyproject.toml
[tool.poetry]
name = "cdk-python-lambda-layers"
version = "0.1.0"
description = ""
authors = ["Albert Asratyan <asratyan.albert@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
# only runtime lambda libraries
python = "^3.12"
aws-lambda-powertools = "^1.28.0"

[tool.poetry.group.dev.dependencies]
# keep all infra libraries + test libraries here
aws-cdk-lib = "^2.162.1"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
